Refactor parse methods from the FlowService into their respectively service classes
